frontend/requirements.txt
pydantic==2.11.3
requests==2.32.5
gradio==5.38.1
python-dotenv==1.2.1
pydantic-settings>=2.0.0